+++ main/ad2_cli_cmd.cpp
@@ -404,6 +404,127 @@ static void _cli_cmd_factory_reset_event(const char *string)
 
 
 /**
+ * @brief Dump report of current tasks sorted by lowest stack free watermark.
+ * FIXME: Move this someplace else? Not exactly HW related. This is OS layer(FreeRTOS).
+ */
+void _pretty_process_list()
+{
+    // Format string for data columns.
+    #define TABBED_LINE_FMT "%-16s%-6s%-9s%-6s%-4s%-4s%-12s%-4s\r\n"
+
+    char buf[1024];
+    std::map<std::string, std::vector<string>> table;
+    std::string recordStr;
+
+    // clear screen home cursor
+    ad2_printf_host(false, "\033[H\033[2J\033[3J");
+
+    // get task list and load into storage container
+    vTaskList(buf);
+    std::stringstream sstreamA(buf);
+
+    std::string processKey;
+    while (std::getline(sstreamA, recordStr)) {
+        std::vector<string> items;
+        istringstream record(recordStr);
+        std::string field;
+        int fi = 0;
+        processKey = "";
+        while (std::getline(record, field, '\t')) {
+            ad2_trim(field);
+            if (field.length()) {
+                fi++;
+                if (fi == 1) {
+                    processKey = field;
+                } else {
+                    items.push_back(field);
+                }
+            }
+        }
+        table[processKey] = items;
+    }
+
+    // get Runtime Stats list and merge into container
+    vTaskGetRunTimeStats(buf);
+    std::stringstream sstreamB(buf);
+    while (std::getline(sstreamB, recordStr)) {
+        istringstream record(recordStr);
+        std::string field;
+        int fi = 0;
+        int tmpA = 0;
+        processKey = "";
+        while (std::getline(record, field, '\t')) {
+            // FIXME: better tokenizer looking for any white space not just \t. Some columns have '\t\t'
+            ad2_trim(field);
+            if (field.length()) {
+                fi++;
+
+                if (fi == 1) {
+                    processKey = field;
+                } else {
+                    if (processKey.length() && table.find(processKey) != table.end()) {
+                        tmpA = table[processKey].size();
+                        table[processKey].push_back(field);
+                    }
+                }
+            }
+        }
+    }
+
+    ad2_printf_host(false, TABBED_LINE_FMT, "Name", "State", "Priority", "Stack", "Num", "CPU", "CNT", "BUSY");
+    for (auto const& x : table)
+    {
+        // sanity check
+        if (x.second.size() == 7) {
+            // format and print the process and details
+            std::string name = x.first.c_str();
+            std::string state = x.second[0];
+            std::string priority = x.second[1];
+            std::string stack = x.second[2];
+            std::string num = x.second[3];
+            std::string cpu = x.second[4];
+            std::string cyc = x.second[5];
+            std::string busy = x.second[6];
+            ad2_printf_host(false, TABBED_LINE_FMT, name.c_str(), state.c_str(), priority.c_str(), stack.c_str(), num.c_str(), cpu.c_str(), cyc.c_str(), busy.c_str());
+        }
+    }
+
+   ad2_printf_host(false,
+        "   State\r\n"
+        "    'B' - Blocked\r\n"
+        "    'R' - Ready\r\n"
+        "    'D' - Deleted (waiting clean up)\r\n"
+        "    'S' - Suspended, or Blocked without a timeout\r\n"
+    );
+
+}
+
+/**
+ * @brief event handler for factory-reset command
+ *
+ * @param [in]string command buffer pointer.
+ *
+ */
+static void _cli_cmd_top_event(const char *string)
+{
+    uint8_t rx_buffer[AD2_UART_RX_BUFF_SIZE];
+    while(1) {
+        _pretty_process_list();
+        vTaskDelay(1000 / portTICK_PERIOD_MS);
+        // Host to AD2*
+        int len = uart_read_bytes(UART_NUM_0, rx_buffer, AD2_UART_RX_BUFF_SIZE - 1, 5 / portTICK_PERIOD_MS);
+        if (len == -1) {
+            // An error happend. Sleep for a bit and try again?
+            ESP_LOGE(TAG, "Error reading for UART aborting task.");
+            break;
+        }
+        if (len>0) {
+            break;
+        }
+    }
+}
+
+/**
  * @brief event handler for netmode command
  *
  * @param [in]string command buffer pointer.
@@ -841,6 +962,14 @@ static struct cli_command cmd_list[] = {
         "\r\n"
         "    Erase config storage and reboot to factory defaults\r\n"
         , _cli_cmd_factory_reset_event
+    },
+    {
+        (char*)AD2_CMD_TOP,(char*)
+        "Usage: top"
+        "\r\n"
+        "    Show process list and stats updated every second.\r\n"
+        "    Press any key to exit.\r\n"
+        , _cli_cmd_top_event
     }
 };