+++ main/ad2_cli_cmd.h
@@ -34,7 +34,8 @@
 #define AD2_CMD_CONFIG   "ad2config"
 #define AD2_CMD_TERM     "ad2term"
 #define AD2_CMD_LOGMODE  "logmode"
-#define AD2_CMD_FACTORY "factory-reset"
+#define AD2_CMD_FACTORY  "factory-reset"
+#define AD2_CMD_TOP      "top"
 #define AD2_CMD_UNLOCKER "teaser :c)"
 
 void register_ad2_cli_cmd(void);